Unify unit tests across history types
