--- packages/hash/tests/unit/global.d.ts
@@ -1,6 +1,0 @@
-declare module NodeJS {
-  interface Global {
-    window: any;
-    document: any;
-  }
-}
