+++ packages/hash/tests/unit/hash.spec.ts
@@ -2,26 +2,11 @@ import "jest";
 import { Hash } from "../../src";
 
 import { withDOM } from "../../../../tests/utils/dom";
+import { navigateSuite } from "../../../../tests/cases/navigate";
 
 // We create our own jsdom instead of using the one that Jest will create
 // so that we can reset the DOM between tests
 describe("Hash constructor", () => {
-  it("returns object with expected API", () => {
-    withDOM({ url: "http://example.com/#/one" }, ({ window }) => {
-      const testHistory = Hash();
-      const expectedProperties = [
-        "location",
-        "action",
-        "toHref",
-        "respondWith",
-        "destroy"
-      ];
-      expectedProperties.forEach(property => {
-        expect(testHistory.hasOwnProperty(property)).toBe(true);
-      });
-    });
-  });
-
   it("initializes using window.location", () => {
     withDOM({ url: "http://example.com/#/one" }, () => {
       const testHistory = Hash();
@@ -130,6 +115,72 @@ describe("Hash constructor", () => {
         const cleanHistory = Hash({ hashType: "clean" });
         expect(cleanHistory.location).toMatchObject({
           pathname: "/the-path"
+        });
+      });
+    });
+  });
+});
+
+describe("navigate()", () => {
+  navigateSuite.forEach(test => {
+    it(test.msg, () => {
+      withDOM({ url: "http://example.com/#/one" }, ({ window }) => {
+        const testHistory = Hash();
+        test.fn({
+          history: testHistory
+        });
+      });
+    });
+  });
+});
+
+describe("toHref", () => {
+  it("returns the location formatted as a string", () => {
+    withDOM({ url: "http://example.com/#/one" }, ({ window }) => {
+      const testHistory = Hash();
+      const currentPath = testHistory.toHref({
+        pathname: "/one",
+        query: "test=query"
+      });
+      expect(currentPath).toBe("#/one?test=query");
+    });
+  });
+
+  const location = { pathname: "/simple-path" };
+
+  describe("hashType", () => {
+    describe("[none provided]", () => {
+      it("outputs expected string", () => {
+        withDOM({ url: "http://example.com/#/one" }, ({ window }) => {
+          const noTypeHistory = Hash();
+          expect(noTypeHistory.toHref(location)).toBe("#/simple-path");
+        });
+      });
+    });
+
+    describe("default", () => {
+      it("outputs expected string", () => {
+        withDOM({ url: "http://example.com/#/one" }, ({ window }) => {
+          const defaultHistory = Hash({ hashType: "default" });
+          expect(defaultHistory.toHref(location)).toBe("#/simple-path");
+        });
+      });
+    });
+
+    describe("bang", () => {
+      it("outputs expected string", () => {
+        withDOM({ url: "http://example.com/#/one" }, ({ window }) => {
+          const bangHistory = Hash({ hashType: "bang" });
+          expect(bangHistory.toHref(location)).toBe("#!/simple-path");
+        });
+      });
+    });
+
+    describe("clean", () => {
+      it("outputs expected string", () => {
+        withDOM({ url: "http://example.com/#/one" }, ({ window }) => {
+          const cleanHistory = Hash({ hashType: "clean" });
+          expect(cleanHistory.toHref(location)).toBe("#simple-path");
         });
       });
     });
