+++ tests/cases/navigate/finish-replace-sets-action.ts
@@ -1,0 +1,13 @@
+import { TestCaseArgs } from "../../types";
+
+export default {
+  msg: "calling pending.finish() sets history's action",
+  fn: function({ history }: TestCaseArgs) {
+    function router(pending) {
+      pending.finish();
+    }
+    history.respondWith(router); // calls router
+    history.navigate("/one");
+    expect(history.action).toBe("replace");
+  }
+};
