+++ tests/cases/navigate/navigate-no-method-same-location.ts
@@ -1,0 +1,15 @@
+import { ignoreFirstCall } from "../../utils/ignoreFirst";
+
+import { TestCaseArgs } from "../../types";
+
+export default {
+  msg: "navigate with no method replaces for same location",
+  fn: function({ history }: TestCaseArgs) {
+    const router = ignoreFirstCall(({ action }) => {
+      expect(action).toBe("replace");
+    });
+    history.respondWith(router);
+
+    history.navigate("/one");
+  }
+};
