+++ tests/cases/navigate/no-response-handler.ts
@@ -1,0 +1,9 @@
+import { TestCaseArgs } from "../../types";
+
+export default {
+  msg: "does nothing without a response handler",
+  fn: function({ history }: TestCaseArgs) {
+    history.navigate("/two");
+    expect(history.location.pathname).toBe("/one");
+  }
+};
