+++ tests/cases/navigate/template.ts
@@ -1,0 +1,6 @@
+import { TestCaseArgs } from "../../types";
+
+export default {
+  msg: "",
+  fn: function({ history }: TestCaseArgs) {}
+};
