+++ tests/utils/ignoreFirst.ts
@@ -1,0 +1,10 @@
+export function ignoreFirstCall(fn) {
+  let notCalled = true;
+  return function() {
+    if (notCalled) {
+      notCalled = false;
+      return;
+    }
+    fn.apply(null, arguments);
+  };
+}
